+++ studio/package.json
@@ -15,7 +15,7 @@
     "@apollo/client": "^3.4.16",
     "@vue/apollo-composable": "^4.0.0-alpha.15",
     "@vueuse/core": "^6.7.3",
-    "ant-design-vue": "^3.2.15",
+    "ant-design-vue": "4",
     "autoprefixer": "^10.3.7",
     "graphql": "^15.6.1",
     "graphql-tag": "^2.12.5",
