+++ studio/src/App.vue
@@ -2,7 +2,6 @@
 import { apolloClient } from "apollo";
 import { provide } from "@vue/runtime-core";
 import Authorized from "components/Authorized.vue";
-import Footer from "components/Footer.vue";
 import Sidebar from "components/Sidebar.vue";
 import { ref } from "vue";
 import { SelectedMenu } from "symbols";
@@ -16,14 +15,30 @@ provide(SelectedMenu, selectedMenu);
 </script>
 
 <template>
-  <Authorized>
-    <a-layout>
-      <Sidebar>
-        <a-layout class="h-screen space-y-2 p-2">
-          <Layout />
-          <Footer />
-        </a-layout>
-      </Sidebar>
-    </a-layout>
-  </Authorized>
+  <a-config-provider
+    :theme="{
+      token: {
+        colorPrimary: '#007011',
+        borderRadius: 4,
+        fontSize: 16,
+        fontFamily: 'Josefin Sans, sans-serif',
+        colorPrimaryBg: '#ffffff',
+      },
+      components: {
+        Footer: {
+          margin: 4,
+        },
+      },
+    }"
+  >
+    <Authorized>
+      <a-layout>
+        <Sidebar>
+          <a-layout class="h-screen space-y-2 p-2">
+            <Layout />
+          </a-layout>
+        </Sidebar>
+      </a-layout>
+    </Authorized>
+  </a-config-provider>
 </template>
