+++ studio/src/components/Sidebar.vue
@@ -5,6 +5,8 @@ import { useQuery } from "@vue/apollo-composable";
 import { StudioGraph } from "../models/studio";
 import gql from "graphql-tag";
 import configs from "../config";
+import StudioVersion from "./StudioVersion.vue";
+import logo from "assets/upstage.png";
 
 const selectedKeys = inject<string[]>(SelectedMenu);
 const iframeSrc = ref("");
@@ -36,12 +38,13 @@ provide(IsAdmin, isAdmin);
 </script>
 
 <template>
-  <a-layout-sider class="bg-transparent">
+  <a-layout-sider theme="light" collapsible class="select-none" width="240">
     <a-spin :spinning="loading">
       <a-menu
         v-model:selectedKeys="selectedKeys"
         mode="inline"
         class="upstage-menu"
+        :open-keys="['admin']"
       >
         <a-menu-item key="media" @click="iframeSrc = ''">
           <picture-outlined />&nbsp;
@@ -58,16 +61,34 @@ provide(IsAdmin, isAdmin);
           <user-outlined />
           <span>Profile</span>
         </a-menu-item>
-        <a-sub-menu key="admin" @click="iframeSrc = ''">
+        <a-sub-menu key="admin">
           <template #icon>
             <key-outlined />
           </template>
           <template #title>Admin</template>
-          <a-menu-item key="admin">Player Management</a-menu-item>
-          <a-menu-item key="6">Batch User Creation</a-menu-item>
-          <a-menu-item key="7">Foyer Customisation</a-menu-item>
-          <a-menu-item key="8">Email Notification</a-menu-item>
-          <a-menu-item key="9">System Configuration</a-menu-item>
+          <a-menu-item key="admin/player" @click="iframeSrc = ''"
+            >Player Management</a-menu-item
+          >
+          <a-menu-item
+            key="admin/batch"
+            @click="iframeSrc = '/backstage/admin/batch-user-creation'"
+            >Batch User Creation</a-menu-item
+          >
+          <a-menu-item
+            key="admin/foyer"
+            @click="iframeSrc = '/backstage/admin/foyer-customisation'"
+            >Foyer Customisation</a-menu-item
+          >
+          <a-menu-item
+            key="admin/email"
+            @click="iframeSrc = '/backstage/admin/email-notification'"
+            >Email Notification</a-menu-item
+          >
+          <a-menu-item
+            key="admin/system"
+            @click="iframeSrc = '/backstage/admin/system-configuration'"
+            >System Configuration</a-menu-item
+          >
         </a-sub-menu>
         <a-menu-item
           key="manual"
@@ -82,32 +103,3 @@ provide(IsAdmin, isAdmin);
   </a-layout-sider>
   <slot></slot>
 </template>
-
-<style lang="less">
-.upstage-menu {
-  margin-top: 8px;
-  margin-left: 8px;
-  margin-right: 0;
-  border-radius: 12px;
-
-  .ant-menu-item {
-    margin: 0;
-    border-radius: 12px;
-
-    &:active {
-      background-color: transparent;
-    }
-
-    &.ant-menu-item-selected {
-      background-color: #147d20;
-      color: white;
-    }
-  }
-}
-
-.ant-tooltip-inner {
-  [role="img"].anticon {
-    margin-right: 8px;
-  }
-}
-</style>
