--- studio/src/styles/custom.less
@@ -1,28 +1,0 @@
-.ant-popover-inner-content {
-  padding: 0;
-}
-.ant-popconfirm {
-  .ant-popover-inner-content {
-    padding: 12px 16px;
-  }
-}
-.ant-table {
-  border-radius: 0.375rem;
-}
-.mb-0 {
-  margin-bottom: 0 !important;
-}
-.ant-input-group.ant-input-group-compact {
-  display: flex;
-}
-.ant-select-selector {
-  height: 100% !important;
-}
-.ant-input-group.ant-input-group-compact::before, .ant-input-group.ant-input-group-compact::after {
-  content: none;
-}
-.ant-tag-has-color:not([style]) {
-  color: rgba(0, 0, 0, 0.85);
-  background: #fafafa;
-  border: 1px solid #d9d9d9;
-}
