+++ studio/src/styles/studio.less
@@ -1,7 +1,6 @@
 // UpStage default font
 @import (css) url('https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100;0,400;0,700;1,400&display=swap');
-@import 'ant-design-vue/lib/notification/style/index.less';
-@import 'ant-design-vue/lib/message/style/index.less';
+@import 'ant-design-vue/dist/reset.css';
 
 // TailwindCSS
 @tailwind components;
