--- studio/src/views/admin/player-management.vue
@@ -1,5 +1,0 @@
-<script setup lang="ts"></script>
-
-<template>
-  <PlayerTable />
-</template>
