+++ studio/src/views/admin/player-management/index.vue
@@ -1,0 +1,11 @@
+<script lang="ts">
+import PlayerFilter from "./PlayerFilter.vue";
+import PlayerTable from "./PlayerTable.vue";
+import { h } from "vue";
+
+export default {
+  setup(props, ctx) {
+    return () => [h(PlayerFilter), h(PlayerTable)];
+  },
+};
+</script>
